qdio: Split SBAL entry flags
Patchwork-id: 52872
O-Subject: [RHEL6.4 PATCH 1/3] [s390x]qdio: Split SBAL entry flags
Bugzilla: 823016
RH-Acked-by: Don Zickus <dzickus@redhat.com>
RH-Acked-by: Jes Sorensen <Jes.Sorensen@redhat.com>

Description
-----------
commit 3ec90878bade9280dee87c9e27d759f1cee07e70
Author: Jan Glauber <jang@linux.vnet.ibm.com>
Commit: Martin Schwidefsky <schwidefsky@de.ibm.com>

    [S390] qdio: Split SBAL entry flags

    The qdio SBAL entry flag is made-up of four different values that are
    independent of one another. Some of the bits are reserved by the
    hardware and should not be changed by qdio. Currently all four values
    are overwritten since the SBAL entry flag is defined as an u32.

    Split the SBAL entry flag into four u8's as defined by the hardware
    and don't touch the reserved bits.


Bugzilla
--------
BZ 823016
https://bugzilla.redhat.com/show_bug.cgi?id=823016

Upstream status of the patch
----------------------------
The patch is upstream as of kernel version 3.0
http://git.kernel.org/?p=linux/kernel/git/torvalds/linux-2.6.git;a=commitdiff;h=3ec90878bade9280dee87c9e27d759f1cee07e70

Test status
-----------
The patch has been tested and fixes the problem.
The fix has been verified by the IBM test department.

+++ drivers/s390/scsi/zfcp_ext.h
@@ -150,7 +150,7 @@ extern struct qdio_buffer_element
 extern struct qdio_buffer_element
 	*zfcp_qdio_sbale_curr(struct zfcp_qdio *, struct zfcp_queue_req *);
 extern int zfcp_qdio_sbals_from_sg(struct zfcp_qdio *,
-				   struct zfcp_queue_req *, unsigned long,
+				   struct zfcp_queue_req *, u8,
 				   struct scatterlist *, int);
 extern int zfcp_qdio_open(struct zfcp_qdio *);
 extern void zfcp_qdio_close(struct zfcp_qdio *);
